requirements.txt
aiohttp==3.8.5python-dotenv==1.0.1setuptools==74.1.2validators==0.34.0wheel==0.44.0